Task_SqlInjection/src/db_init.py
import os
import psycopg2
from dotenv import load_dotenv

load_dotenv()
DATABASE_URL = os.getenv("DATABASE_URL")

def init_db():
    conn = psycopg2.connect(DATABASE_URL)
    cur = conn.cursor()
    cur.execute("CREATE TABLE IF NOT EXISTS users (id SERIAL PRIMARY KEY, username TEXT, secret TEXT);")
    conn.commit()
    cur.close()
    conn.close()

if __name__ == "__main__":
    init_db()
