Task_SqlInjection/src/requirements.txt
fastapi
uvicorn
python-dotenv
psycopg2-binary
pycryptodome
pydantic
